feat: add StatCard component for displaying statistics on the homepage
